src/python_clip/pyproject.toml
[project]
name = "python-clip"
version = "0.1.0"
dependencies = [
    "torch>=2.6.0",
    "transformers>=4.48.3",
    "pillow>=11.1.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0"
]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
